+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * "This semester" shortcut
 
+### Changed
+* Improve feedback when importing subjects
+
 ### Fixed
 * Subject picker too large
 * Empty name bug for subjects and categories